add insulin counterfactual metrics for dose and timing analysis
