+++ synthetic_data/data/dose_counterfactuals/insulin_counterfactual_metrics.csv
@@ -1,0 +1,6 @@
+insulin_factor,mean_glucose,std_glucose,min_glucose,max_glucose,time_in_range,time_below_range,time_above_range
+1.0,108.26636904761905,23.451507010271605,49,193,98.01587301587301,1.4384920634920635,0.5456349206349206
+0.8,117.20287698412699,29.46340644362561,61,209,97.22222222222221,0.6944444444444444,2.083333333333333
+0.9,112.94047619047619,26.01412917428156,55,200,98.16468253968253,1.0912698412698412,0.744047619047619
+1.1,103.82093253968254,21.469493119594613,41,184,97.0734126984127,2.6785714285714284,0.248015873015873
+1.2,99.45882936507937,20.639918715544127,40,177,95.13888888888889,4.861111111111112,0.0
