+++ synthetic_data/data/timing_counterfactuals/insulin_timing_counterfactual_metrics.csv
@@ -1,0 +1,6 @@
+timing_shift_min,mean_glucose,std_glucose,min_glucose,max_glucose,time_in_range,time_below_range,time_above_range
+0,108.26636904761905,23.451507010271605,49,193,98.01587301587301,1.4384920634920635,0.5456349206349206
+-60,114.43700396825396,49.73882157794118,40,243,67.85714285714286,16.964285714285715,15.178571428571427
+-30,109.6046626984127,37.49431001444582,40,221,83.28373015873017,11.011904761904761,5.704365079365079
+30,108.42013888888889,30.805619209310464,72,220,94.79166666666666,0.0,5.208333333333334
+60,108.4250992063492,49.78902731209902,40,249,71.97420634920636,14.037698412698413,13.988095238095239
